+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+**?.?.? Changes:**
+
+* Added damage display for Heretic Ruin.
+
 **1.3.6 Changes:**
 
 * Fixed damage calculations for Warped Echo and Breaching Fin changed in SotS phase 2 patch.